+++ htmlcov/index.html
@@ -17,7 +17,7 @@
 <div id="header">
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">100.00%</span>
+            <span class="pc_cov">3.74%</span>
         </h1>
         <img id="keyboard_icon" src="keybd_closed.png" alt="Show keyboard shortcuts" />
         <form id="filter_container">
@@ -56,23 +56,32 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>1</td>
+                <td>79</td>
+                <td>75</td>
                 <td>0</td>
+                <td>28</td>
                 <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="1 1">100.00%</td>
+                <td class="right" data-ratio="4 107">3.74%</td>
             </tr>
         </tfoot>
         <tbody>
             <tr class="file">
                 <td class="name left"><a href="src_oemof_solph___init___py.html">src/oemof/solph/__init__.py</a></td>
-                <td>1</td>
+                <td>7</td>
+                <td>5</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
+                <td class="right" data-ratio="2 7">28.57%</td>
+            </tr>
+            <tr class="file">
+                <td class="name left"><a href="src_oemof_solph_network_py.html">src/oemof/solph/network.py</a></td>
+                <td>72</td>
+                <td>70</td>
                 <td>0</td>
-                <td class="right" data-ratio="1 1">100.00%</td>
+                <td>28</td>
+                <td>0</td>
+                <td class="right" data-ratio="2 100">2.00%</td>
             </tr>
         </tbody>
     </table>
@@ -84,7 +93,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io">coverage.py v5.0.4</a>,
-            created at 2020-03-25 13:51
+            created at 2020-03-26 14:29
         </p>
     </div>
 </div>
