+++ htmlcov/src_oemof_solph___init___py.html
@@ -3,7 +3,7 @@
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=emulateIE7" />
-    <title>Coverage for src/oemof/solph/__init__.py: 100.00%</title>
+    <title>Coverage for src/oemof/solph/__init__.py: 28.57%</title>
     <link rel="stylesheet" href="style.css" type="text/css">
     <script type="text/javascript" src="jquery.min.js"></script>
     <script type="text/javascript" src="jquery.hotkeys.js"></script>
@@ -17,13 +17,13 @@
 <div id="header">
     <div class="content">
         <h1>Coverage for <b>src/oemof/solph/__init__.py</b> :
-            <span class="pc_cov">100.00%</span>
+            <span class="pc_cov">28.57%</span>
         </h1>
         <img id="keyboard_icon" src="keybd_closed.png" alt="Show keyboard shortcuts" />
         <h2 class="stats">
-            1 statements &nbsp;
-            <span class="run shortkey_r button_toggle_run">1 run</span>
-            <span class="mis show_mis shortkey_m button_toggle_mis">0 missing</span>
+            7 statements &nbsp;
+            <span class="run shortkey_r button_toggle_run">2 run</span>
+            <span class="mis show_mis shortkey_m button_toggle_mis">5 missing</span>
             <span class="exc show_exc shortkey_x button_toggle_exc">0 excluded</span>
                 <span class="par run show_par shortkey_p button_toggle_par">0 partial</span>
         </h2>
@@ -53,12 +53,21 @@
 </div>
 <div id="source">
     <p id="t1" class="run"><span class="n"><a href="#t1">1</a></span><span class="t"><span class="nam">__version__</span> <span class="op">=</span> <span class="str">"0.4.0.dev0"</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t2" class="pln"><span class="n"><a href="#t2">2</a></span><span class="t">&nbsp;</span><span class="r"></span></p>
+    <p id="t3" class="run"><span class="n"><a href="#t3">3</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">network</span> <span class="key">import</span> <span class="op">(</span><span class="nam">Sink</span><span class="op">,</span> <span class="nam">Source</span><span class="op">,</span> <span class="nam">Transformer</span><span class="op">,</span> <span class="nam">Bus</span><span class="op">,</span> <span class="nam">Flow</span><span class="op">,</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t4" class="pln"><span class="n"><a href="#t4">4</a></span><span class="t">                      <span class="nam">EnergySystem</span><span class="op">)</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t5" class="mis show_mis"><span class="n"><a href="#t5">5</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">models</span> <span class="key">import</span> <span class="nam">Model</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t6" class="mis show_mis"><span class="n"><a href="#t6">6</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">groupings</span> <span class="key">import</span> <span class="nam">GROUPINGS</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t7" class="mis show_mis"><span class="n"><a href="#t7">7</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">options</span> <span class="key">import</span> <span class="nam">Investment</span><span class="op">,</span> <span class="nam">NonConvex</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t8" class="mis show_mis"><span class="n"><a href="#t8">8</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">plumbing</span> <span class="key">import</span> <span class="nam">sequence</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t9" class="mis show_mis"><span class="n"><a href="#t9">9</a></span><span class="t"><span class="key">from</span> <span class="op">.</span><span class="nam">components</span> <span class="key">import</span> <span class="op">(</span><span class="nam">GenericStorage</span><span class="op">,</span> <span class="nam">GenericCHP</span><span class="op">,</span> <span class="nam">ExtractionTurbineCHP</span><span class="op">,</span>&nbsp;</span><span class="r"></span></p>
+    <p id="t10" class="pln"><span class="n"><a href="#t10">10</a></span><span class="t">                         <span class="nam">OffsetTransformer</span><span class="op">)</span>&nbsp;</span><span class="r"></span></p>
 </div>
 <div id="footer">
     <div class="content">
         <p>
             <a class="nav" href="index.html">&#xab; index</a> &nbsp; &nbsp; <a class="nav" href="https://coverage.readthedocs.io">coverage.py v5.0.4</a>,
-            created at 2020-03-25 13:48
+            created at 2020-03-26 14:29
         </p>
     </div>
 </div>
