+++ htmlcov/status.json
@@ -1,1 +1,1 @@
-{"format":2,"version":"5.0.4","globals":"d156ef72399fd244292f4bd8b44cdd27","files":{"src_oemof_solph___init___py":{"hash":"30a727461e8b6caa48f66b5a7bea896a","index":{"nums":[1,1,0,0,0,0,0],"html_filename":"src_oemof_solph___init___py.html","relative_filename":"src/oemof/solph/__init__.py"}}}}
+{"format":2,"version":"5.0.4","globals":"1a9b5388313a061a2515bdfe71f533d3","files":{"src_oemof_solph___init___py":{"hash":"d2a3b411838219cbc6e231c81e24f8d1","index":{"nums":[1,7,0,5,0,0,0],"html_filename":"src_oemof_solph___init___py.html","relative_filename":"src/oemof/solph/__init__.py"}},"src_oemof_solph_network_py":{"hash":"8be5bb2c174a78b251ff7afb006e348c","index":{"nums":[1,72,0,70,28,0,28],"html_filename":"src_oemof_solph_network_py.html","relative_filename":"src/oemof/solph/network.py"}}}}
