--- oemof/__init__.py
@@ -1,4 +1,0 @@
-from pkgutil import extend_path
-__path__ = extend_path(__path__, __name__)
-
-__version__ = '0.4.0dev'
