+++ src/oemof.solph.egg-info/SOURCES.txt
@@ -1,9 +1,4 @@
-.appveyor.yml
-.bumpversion.cfg
-.cookiecutterrc
 .coveragerc
-.editorconfig
-.readthedocs.yml
 .travis.yml
 AUTHORS.rst
 CHANGELOG.rst
@@ -20,22 +15,155 @@ ci/requirements.txt
 ci/templates/.appveyor.yml
 ci/templates/.travis.yml
 ci/templates/tox.ini
+docs/about_oemof.rst
+docs/api.rst
 docs/authors.rst
 docs/changelog.rst
 docs/conf.py
 docs/contributing.rst
+docs/developing_oemof.rst
+docs/docutils.conf
+docs/getting_started.rst
 docs/index.rst
 docs/installation.rst
+docs/installation_and_setup.rst
+docs/oemof_network.rst
+docs/oemof_outputlib.rst
+docs/oemof_solph.rst
+docs/oemof_tools.rst
 docs/readme.rst
 docs/requirements.txt
 docs/spelling_wordlist.txt
 docs/usage.rst
+docs/using_oemof.rst
+docs/whats_new.rst
+docs/_files/pahesmf_init.py
+docs/api/modules.rst
+docs/api/oemof.outputlib.rst
+docs/api/oemof.rst
+docs/api/oemof.solph.rst
+docs/api/oemof.tools.rst
 docs/reference/index.rst
 docs/reference/oemof.solph.rst
-src/oemof.solph.egg-info/PKG-INFO
-src/oemof.solph.egg-info/SOURCES.txt
-src/oemof.solph.egg-info/dependency_links.txt
-src/oemof.solph.egg-info/not-zip-safe
-src/oemof.solph.egg-info/top_level.txt
-src/oemof/solph/__init__.py
-tests/test_oemof_solph.py
+docs/whatsnew/v0-0-1.rst
+docs/whatsnew/v0-0-2.rst
+docs/whatsnew/v0-0-3.rst
+docs/whatsnew/v0-0-4.rst
+docs/whatsnew/v0-0-5.rst
+docs/whatsnew/v0-0-6.rst
+docs/whatsnew/v0-0-7.rst
+docs/whatsnew/v0-1-0.rst
+docs/whatsnew/v0-1-1.rst
+docs/whatsnew/v0-1-2.rst
+docs/whatsnew/v0-1-4.rst
+docs/whatsnew/v0-2-0.rst
+docs/whatsnew/v0-2-1.rst
+docs/whatsnew/v0-2-2.rst
+docs/whatsnew/v0-2-3.rst
+docs/whatsnew/v0-3-0.rst
+docs/whatsnew/v0-3-1.rst
+docs/whatsnew/v0-3-2.rst
+docs/whatsnew/v0-3-3.rst
+docs/whatsnew/v0-4-0.rst
+oemof.solph.egg-info/PKG-INFO
+oemof.solph.egg-info/SOURCES.txt
+oemof.solph.egg-info/dependency_links.txt
+oemof.solph.egg-info/entry_points.txt
+oemof.solph.egg-info/not-zip-safe
+oemof.solph.egg-info/requires.txt
+oemof.solph.egg-info/top_level.txt
+oemof.solph/__init__.py
+oemof.solph/blocks.py
+oemof.solph/components.py
+oemof.solph/constraints.py
+oemof.solph/custom.py
+oemof.solph/groupings.py
+oemof.solph/models.py
+oemof.solph/network.py
+oemof.solph/options.py
+oemof.solph/plumbing.py
+oemof.solph/processing.py
+oemof.tools/__init__.py
+oemof.tools/console_scripts.py
+oemof.tools/debugging.py
+oemof.tools/economics.py
+oemof.tools/helpers.py
+oemof.tools/logger.py
+tests/basic_tests.py
+tests/constraint_tests.py
+tests/regression_tests.py
+tests/run_nose.py
+tests/solph_tests.py
+tests/test_components.py
+tests/test_console_scripts.py
+tests/test_groupings.py
+tests/test_models.py
+tests/test_network_classes.py
+tests/test_processing.py
+tests/test_solph_network_classes.py
+tests/test_warnings.py
+tests/tool_tests.py
+tests/lp_files/activity_costs.lp
+tests/lp_files/connect_investment.lp
+tests/lp_files/dsm_module_delay.lp
+tests/lp_files/dsm_module_interval.lp
+tests/lp_files/emission_limit.lp
+tests/lp_files/fixed_source_invest_sink.lp
+tests/lp_files/fixed_source_variable_sink.lp
+tests/lp_files/flow_invest_with_offset.lp
+tests/lp_files/flow_invest_with_offset_no_minimum.lp
+tests/lp_files/flow_invest_without_offset.lp
+tests/lp_files/invest_source_fixed_sink.lp
+tests/lp_files/investment_limit.lp
+tests/lp_files/linear_transformer.lp
+tests/lp_files/linear_transformer_chp.lp
+tests/lp_files/linear_transformer_chp_invest.lp
+tests/lp_files/linear_transformer_invest.lp
+tests/lp_files/max_source_min_sink.lp
+tests/lp_files/maximum_shutdowns.lp
+tests/lp_files/maximum_startups.lp
+tests/lp_files/min_max_runtime.lp
+tests/lp_files/nominal_value_to_zero.lp
+tests/lp_files/offsettransformer.lp
+tests/lp_files/source_with_gradient.lp
+tests/lp_files/storage.lp
+tests/lp_files/storage_fixed_losses.lp
+tests/lp_files/storage_invest_1.lp
+tests/lp_files/storage_invest_1_fixed_losses.lp
+tests/lp_files/storage_invest_2.lp
+tests/lp_files/storage_invest_3.lp
+tests/lp_files/storage_invest_4.lp
+tests/lp_files/storage_invest_5.lp
+tests/lp_files/storage_invest_6.lp
+tests/lp_files/storage_invest_all_nonconvex.lp
+tests/lp_files/storage_invest_minimum.lp
+tests/lp_files/storage_invest_unbalanced.lp
+tests/lp_files/storage_invest_with_offset.lp
+tests/lp_files/storage_invest_without_offset.lp
+tests/lp_files/storage_unbalanced.lp
+tests/lp_files/transformer.lp
+tests/lp_files/transformer_invest.lp
+tests/lp_files/transformer_invest_with_existing.lp
+tests/lp_files/variable_chp.lp
+tests/test_outputlib/__init__.py
+tests/test_outputlib/input_data.csv
+tests/test_outputlib/test_views.py
+tests/test_scripts/test_solph/test_connect_invest/connect_invest.csv
+tests/test_scripts/test_solph/test_connect_invest/test_connect_invest.py
+tests/test_scripts/test_solph/test_flexible_modelling/test_add_constraints.py
+tests/test_scripts/test_solph/test_generic_caes/generic_caes.csv
+tests/test_scripts/test_solph/test_generic_caes/test_generic_caes.py
+tests/test_scripts/test_solph/test_generic_chp/ccet.csv
+tests/test_scripts/test_solph/test_generic_chp/test_generic_chp.py
+tests/test_scripts/test_solph/test_lopf/test_lopf.py
+tests/test_scripts/test_solph/test_simple_model/input_data.csv
+tests/test_scripts/test_solph/test_simple_model/test_simple_dispatch.py
+tests/test_scripts/test_solph/test_simple_model/test_simple_dispatch_one.py
+tests/test_scripts/test_solph/test_simple_model/test_simple_invest.py
+tests/test_scripts/test_solph/test_storage_investment/es_dump_test_3_2dev.oemof
+tests/test_scripts/test_solph/test_storage_investment/storage_investment.csv
+tests/test_scripts/test_solph/test_storage_investment/test_invest_storage_regression.py
+tests/test_scripts/test_solph/test_storage_investment/test_storage_investment.py
+tests/test_scripts/test_solph/test_storage_investment/test_storage_with_tuple_label.py
+tests/test_scripts/test_solph/test_variable_chp/test_variable_chp.py
+tests/test_scripts/test_solph/test_variable_chp/variable_chp.csv
